Add complete ecb test code
This commit includes single and multithreaded ecb, though the
multithreaded version is slower than the 1x version for some
reason. I believe it has to do with the redundancy in the
argument passing during nested function calls which I can improve
upon should I be so motivated. All in all this was a good project
as I got to work with basic crypto and got to use new features of
the C language like the pthread library and more expressive
macros for debugging and timing. Clean up is still needed but
good work!