Begin capacity to encrypt input from a file
I would like to take input from a file to give the user
more control. I also think that it would be nice to have a
struct that I can use for all the modes of operation and to make
threading them all easier because I can then pass it into the
create function.