Start blind coding of guessing game
Blind just means that you aren't looking at the book's
instructions but writting the program on your own. You tried to
get the StdRng tuple struct working but the compiler kept
complaining about not being able to create it due to the StdRng
struct having private fields... which it doesn't have. You'll
have to look into this more in the future you only spent 15 min
on it.