Add SLC (smpl lang compiler) directory
This is a backup of what I had written at the end of the quarter.
If stuff hits this fan at the very least I can come back to this.
It lexes the entirety of the inputfile first and then parses it,
which is slower than doing both simultaneously and since this is
an optimizing compilers course I should probably do the latter.