Remove typedefs in favor of explicitly using struct keyword
Also starting to create the AST but you realized that you kind of
already have everything you just haven't connected everything up
should you wrap the things in AstNodes to connect them? Or should
you just add fields to the other types that you've created? I'm
leaning toward wrapping, but it seems like it will mess up the
flow of a lot of the parser functions?